# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlphaAgent API",
    "description": "Financial research agent with tool use, guardrails, and multi-turn chat.",
    "version": "2.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Health",
        "description": "Liveness probe \u2014 also reports backend, model, auth mode and cache status.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Metrics",
        "description": "Prometheus text-format metrics.",
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Metrics Metrics Get"
                }
              }
            }
          }
        }
      }
    },
    "/tools": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Tools",
        "description": "List the tools the agent can use (name + description).",
        "operationId": "tools_tools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sessions": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Create Session",
        "description": "Create a new multi-turn conversation session.",
        "operationId": "create_session_sessions_post",
        "parameters": [
          {
            "name": "_",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": " "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ask": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Ask",
        "description": "Run the guarded agent on a financial question.\n\nMulti-turn: pass ``session_id`` to continue a conversation.  Without\nit, the question is answered standalone.\n\nGuards apply automatically \u2014 prompt-injection / financial disclaimer\n(input/output), ticker validation and tool rate limiting, and\nanti-hallucination grounding of numbers in the answer.",
        "operationId": "ask_ask_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          },
          "backend": {
            "type": "string",
            "title": "Backend",
            "default": "ollama"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "cache_enabled": {
            "type": "boolean",
            "title": "Cache Enabled",
            "default": false
          },
          "auth_mode": {
            "type": "string",
            "title": "Auth Mode",
            "default": "none"
          }
        },
        "type": "object",
        "title": "HealthResponse"
      },
      "QueryRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Conversation session id (optional)"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "grounded": {
            "type": "boolean",
            "title": "Grounded",
            "default": true
          },
          "blocked": {
            "type": "boolean",
            "title": "Blocked",
            "default": false
          },
          "block_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Reason"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolResultInfo"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "iterations_used": {
            "type": "integer",
            "title": "Iterations Used",
            "default": 0
          },
          "from_cache": {
            "type": "boolean",
            "title": "From Cache",
            "default": false
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "QueryResponse"
      },
      "SessionCreateResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "SessionCreateResponse"
      },
      "ToolCallInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ToolCallInfo"
      },
      "ToolResultInfo": {
        "properties": {
          "tool_call": {
            "$ref": "#/components/schemas/ToolCallInfo"
          },
          "output": {
            "type": "string",
            "title": "Output",
            "default": ""
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "tool_call"
        ],
        "title": "ToolResultInfo"
      },
      "ToolsResponse": {
        "properties": {
          "tools": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tools"
          }
        },
        "type": "object",
        "required": [
          "tools"
        ],
        "title": "ToolsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}